refactor screener finviz_view to view and model
